Licensing and log minor fixes
Two improvements:

Now, the LicenseException is caught for ICliAction plugins and a
simplified log message is written out without a stacktrace, because we
know exactly what went wrong. In addition, an easier readable log
message generated for CompoenentSettings, when they throw exceptions.

Second, the log file is renamed earlier and not just inside the 'run'
action, but instead at the action 'executor' level.