4999 Available Values Merged and Input
Issue 4999: When available values got multi-selected or in other ways combined,
selection would not work. This has been fixed by explicitly having a
strategy for MergedValueAnnotations.

Issue 5009: The Input.Value getter threw an exception, which in turn
could cause issues when the member of that was annotated. Specifically
this was due to a problem with the way some values of an enum can be
hidden, and this in turn affected Input<Verdict> for test steps. A
number of protetions were enabled here. BlockMembersAnnotationAttribute
helps the annotator skip the class with the issue (Input<>), Enum
annotations has a checkBrowsable now that is checked before fetching the
value. AccessAnnotationProxy now only loads IsReadOnly and IsVisible
when needed. And DefaultValueAnnotation catches the exception when it
occurs. So, 4 improvements that should improve overall stability of
the annotation system.

Another new feature is the IAvailableValuesSelectedAnnotation, which
provides an interface to set the SelectedValue in the same
AvailableValuesAnnotation. This provides better support for advanced
drop-down behavior, but for now it is only part of the internal API and
not used. Probably will be in the future though.