4742 TestPlan Open Hash
The test plan did not get re-serialized after TestPlan.Open. This meant that the serialized plan did not correspond the the test plan that actually was executed. TestPlan.Open is only meant to optimize by avoiding opening and closing resources all the time, so skipping serialization is not really a part of that optimization.

A unit test was added that ensures that the test plan and hash of the test plan is correct.

Also fixed a minor inconsistency when a result listener threw a OperationCancelledException due to the test plan being aborted.