OpenTAP SDK new project nuget version
The GenerateProject CLI action now learned to get the newest TAP version from the nuget server. Maybe we should restrict it to be the same major version as the one from the local installation? For now its not an issue.

It also can handle if it is unable to connect. In this case it tries get a version matching X.Y.Z to what is currently installed - I believe this should be possible except when using a pre-release.

A change to our process could be to push a nuget package with the newest from the beta branch and use that in case of pre-release, so if the local version is a prerelease, then we just install X.Y.Z-beta, so same version for all betas for the next version of tap. Anyway, just an idea...