5064 sweep loop range clear parameters: Test Failure
A unit test failure caused by the members being added before the steps were added. Since the new behavior is to check that the members are indeed valid, when the list changes, those changes got deleted, hence the failure. This has been fixed in the unit test, by setting the members after setting the child steps.