100 Scpi Instrument Read STB Not Supported
Added check if readstb can be supported, otherwise it uses ERR? to detect errors. Also added IScpiIO2 to make it possible to check that the functionality of the readstb alternative works in a unit test, which was also added. The most tricky parts of IScpiIO2 was to make SRQ events work, it has not been checked if these still work or whether they even worked before. Since its a rarely used API, we decided to wait for someone to discover any bugs that might be present, the functionality should be as before though. Related to SRQ, what was detected was that it probably does not work with SOCKET based instruments and this resulted in an error when the event was attached.