24 MetaData Prompt Order
The PlatformInteractionRace unit test was fixed to add the that TestPlanRunStarted has not happened.

I was only able to reproduce under some circumstances when TestPlan.Open was being used. This has now been fixed. Especially there was an issue with the Lazy resource manager and metadata prompt and this has now been fixed.