135 Timeguard with break conditions
The timeguard step no longer uses CancellationExceptions to cancel the test plan. In addition, it can override the verdict when the timeout occures, the default value of the verdict is Error. It is not 'Aborted', because that would cause the test plan to not be able to continue due to verdict cascading.

- Added a unit test that verifies the functionality.