Implemented general "parameterization" feature.
Any setting/IMemberData can be "parameterized" onto any other object (to create a "parameter" on that object). Parameterizing onto the TestPlan makes it an External Parameter - onto a parent TestStep makes it a "Scoped Parameter".

Two new sweep steps are added that uses parameters. The two existing sweep steps are moved to a "Legacy" group to be hidden completely later.