281 Package Install Overwrite Validation:
Added support for detecting harmful overwrites by checking the hash of installed vs about-to-install packages. This is done before downloading the files, so the situation can be quickly reported back to the user without waiting for downloads.

Simplified DependencyChecker a bit so it does not use Installation and thereby is easier to test.

Also ignoring conflicts found in the Dependencies folder.

Added unit test to verify that it works. Also added unit test to DependencyChecker.