280 Decaying Items in memorizer
the Memorizer class supported timing objects out after a certain amount of time had passes(30s) and it was being used by ComponentSettings. This does not really make sense for component settings, since there are live changes in them that should not be removed. In most cases this is not a problem, because settings are normally saved right after being modified, but if one modifies components settings while the test plan is running without saving them, this issue might occur.

The default value has now been increased to TimeSpan.MaxValue (10675199 days).