Steps which parameterize a resource of a child step should not depend on those resources.
Currently, if a resource is parameterized on a top-level step, it will be open for the entirety of the test plan run, when it is probably only used for a small portion of it. This MR fixes that behavior.
The behavior change mentioned above also happens to fix the particular instance of the problem that this issue refers to, but we are still not checking the EnabledIf property on parameterized steps. Doing this would make the complexity of EnabledIf checking quadratic in the depth of the testplan.

Since EnabledIf can be a performance bottleneck, I also added some performance tests along with the tests to verify the new resource behavior.