542 HybridStreamLogLeak: Fix
Fixed the issue by making the stream backing file auto delete on dispose (or application exit). This flag is honored by Windows, so it automatically gets deleted even if the application crashes.

Added a unit test that tests hybrid log stream, but it does not test if the file is deleted on application crash, that is left to be manually tested. (I verified it works).

Improved the temp file name so it is easier to spot in the temp folder. (.opentap-log.txt)