537 Delete Parameterized Step: Fix
Added a fix, so that the changeid does not get updated when parameter members gets deleted. This has the effect that checkParametersanity will be called until everything is finally 'sane'. And this fixes this issue.

I was not able to create a unit test that verifies the issue, because the issues does not reproduce in simple code. It is very easy to reproduce in the user interface however, so it is probably not a data race.

In any case, the current solution should work.