644 Annotation Read Performance Issue: Fix
Made sure that MemberDataAnnotations are not being read more than
necessesary. This was always the intention, but it seems that it has
been removed. Possibly due to a bug fix. Since there is no hint about
that in the code lets re-enable the cache as is also done when the value
is set.