666 If Verdict Parallel Stuck: fix
Fixed the issue by inversing the order of logic in the loop checking that the output is ready. It now starts from the bottom, ensuring each level has some kind of step run, waiting at each level if necessesary until finally it gets the step run of the child step in question.

It should be relatively fast in the normal case.