Optimize build-time package installs
Use the OpenTAP Image feature to install packages during build instead of installing packages one at a time.

This solution uses reflection and does not start any subprocesses. Previously, if a project installed 10 different packages, it would start a new process for each package. If the specified packages were not compatible, the error message was not very helpful. Now we get exact error messages describing the problem in a way that is much easier to maintain because it must be supported by the Image API.

Although there is a lot of reflection, the reflected methods and properties are all public, so they should be completely stable.