Install packages as an Image during build
Use the OpenTAP Image feature to install packages during build instead of installing packages one at a time.

This solution starts 0 process instead of starting once process for each package to install. It is also much more stable and ensures that the resulting installation is not broken, which was not the case previously.