ISettings Reference Icon Annotation:
Made use of an optimization named IParameterizedMembersCache to optimized reverse-lookup of i member data. Note, that the slow part is not iterating the parent tree, but iterating through all the members of each level of the tree. These may contains 20s of parameters and this needs to be done for each parameter in an O(N^2) fashion.