255 available values update bug (#257)
The issue was related to Read, reading AvaliableValues from other IAvailableValuesAnnotations. This was done before Read() was called on the other object, causing the value read to be invalid. By waiting for the getter to be called, we can insure that all objects has been Read and avoiding that some intermediate state is cached.