119 Incompatible DLL Warning: fix and test (#286)
Added a warning if the --OS and --Architecture is specified but the target installation is not compatible.
Added returning an error if --OS and --Architecture is _not_ specified and the target installation is not compatible.

Added a unit test to verify the behavior. The unit tests behaves differently depending on platform. If on Windows, the normal behavior will be tested but on Linux the incompatible package install situation will be tested.