Remove 'Isolation' logic in favor of loading assemblies from streams
This is part of an effort to centralize assembly loading.
All assembly loading is now being done by the PluginManager itself

When an assembly is loaded from a stream, the 'Assembly.Location' field
is not populated. We rely pretty heavily on this being present,
so when running isolated we now need a lookup to get this location
instead. This is likely going to break some third party plugins
specifically when running isolated, but this shouldn't affect many
people.

A huge amount of subprocess related logic has been removed, and cleanups
were possible in many of the places where I had to replace
'Assembly.Location' usages

Centralizing assembly loading is a nice stepping stone for being able to
install plugins inside of ausing (OpenTap.Session) {/* code */} block
and uninstalling it afterwards,
but more work is needed for this to function correctly, as the
assemblies will still be loaded even after uninstalling.

I measured the new method to be around 1 second faster than the previous
method, and especially when running unsigned alpha versions of OpenTAP,
anti-virus programs should cause less issues with the new method.