359 Reflection Helper Optimization: fix
Added optimization of GetAttributes. The biggest change is avoiding allocation when possible. The path when the list of results is empty is optimized to do no allocation and only return Array.Empty. The result which is exactly one element is also optimized.

Often attributes are array, this case is also optimized by explicitly looping over the array.