Available values re-read issue (#379)
The issue was caused by annotated elements from Generic Sequence Annotation not being reused when possible. Previously, when any value change, the entire list of elements needs to be re-read. Even if just a value changed.

This caused issues since the GUI only refreshes the full list when elements are added or removed, but not when values inside elements are changed. So instead, if we can detect that no changes has been done to the list, we can re-use most of the old annotations.

This also has the benefit of improving the performance, since with dont need to re-generated annotation collections as often.

In turn this caused an issue inside sweepRow (sweep parameter loop). There is a cache inside ObjectValueAnnotation that assumes that the type of an object never changes. We now allow the type of objects inside ObjectValueAnnotations to change when a Read occurs in the AnnotationCollection tree.

Also improved printing of AnnotationCollections.