Read SCPI Block with termination character included in the last char.
A hack was introduced to fix issues with instruments that includes the termination character in the bytes read from a IEEE Block.
In this case we should not wait for the last termination character as that has already been sent and doing so will cause in a timeout exception in the best case.