Bug fix and clean up: (#381)
Removed AssemblyExtensions.GetSemanticVersion which is not supported by .net framework assemblies. Also we don't really need it anymore, since the package version is the more 'right' way to get information about a plugin.

Using TypeData to get the assembly version of the currently executing assembly.

Added a fix for when assemblies does not have a semantic version.