Fix transitive dependency resolution
This fixes a bug causing transitive dependencies to not be correctly resolved when deploying to an existing installation. This bug occurred because the existing installation would enforce that same architecture for all resolved packages. Due to a bug in the package repository, transitive dependencies with the architecture 'AnyCPU' does not satisfy x64. This is not possible to fix on the server side, so OpenTAP must relax the requirements a bit and handle compatibility resolution itself.

This also extends the testability of the Image functionality quite a bit by adding an implementation for 'downloading' mocked packages. This meant that we can greatly speed up our deployment tests because we don't need to communicate with any repositories, and we're deploying very small packages. It also allows us to reenable some tests that were previously disabled.