408 WorkQueue Thread Control
Added support for specifying the thread context for the work queue. That means that job executed in the work queue has a specific thread context instead of whatever the first enqueuer had.

This is needed for situations where we want to keep accepting events even after that a test plan is aborted.

added a unit test to verify the new behavior.