Reasonable get package from repo (#410)
* Use OpenTAP dll from output folder instead of nuget dir

* Fix transitive dependency resolution

This fixes a bug causing transitive dependencies to not be correctly resolved when deploying to an existing installation. This is due to a bug in the package repository's "compatibleWith" dependency satisfaction logic. We will fix this by not relying on the repo to do this, and handle dependency/compatibility resolution completely within OpenTAP.

This also extends the testability of the Image functionality quite a bit by adding an implementation for 'downloading' mocked packages. This meant that we can greatly speed up our deployment tests because we don't need to communicate with any repositories, and we're deploying very small packages. It also allows us to reenable some tests that were previously disabled.

* Refator DependencyResolver.GetPackagesFromRepo to become more readable and not rely on PackageRepository compatibleWith logic

* Unit test to show #386 only runs with the internal repo - disable it for CI.

Co-authored-by: allarsen <alexander.larsen@keysight.com>
Co-authored-by: Asger Iversen <asger_iversen@keysight.com>