541 TypeData Attributes gets unintentionally inherited: Fix
It seems a programming error inside GetAllCustomAttributes had caused the behavior to be flipped between inherit= true and false.

This could cause some unintentional sideeffects since it has been this way for some time. Possibly we should just say the standard is to always inherit, but that limits a lot how attributes could be used.