537 type data unbrowsable bug (#540)
* TypeData Cleanup

Merged all type data partial classes into one and ordered members to make it easier to get an overview.

* 537 TypeDataUnbrowsableBug: Unit test

Added a unit test which shows that TypeData.FromType creates an unbrowsable type in this case. Also added a test that IsBrowsable should be false if Browsable(False) is used.

* 537 TypeDataUnbrowsable bug: Fix

- Fixed the issue by calling the more basic constructor from the (Type)- constructor. This ensures that IsBrowsable is true in default case. Also added a check for BrowsableAttribute to see if IsBrowsable should be false.

- Did a couple of cleanup with regards to the TypeData constructor which in all cases could take a 'name'. -> TypeData.Name can be get-only.

* 537 TypeData Unbrowsable Bug: Test improvement

Added an additional case to test. Now we t est both Browsable(false) and Browsable(true).