765 Many Results Thread Block: Fix
Fixed the issue by waiting(decrementing) on the 'addSemaphore' in the work queue when dequeuing a work item.

This fixes the issue because otherwise we would 'leak' semaphore counts until we reach the 'semaphoreMaxCount' which then forces us to block until some work items has been processed, which never happens because there is no work items left to process.