774 Available Values Overwritten during parameterization: Fix + test
Fixed the issue by changing the way we control list equality.

The annotation system thinks the lists has been changed because the returned value from merged value annotation was not null initially and the value becomes aligned unintentionally afterwards.

This is really a corner case for Browsable(False) properties which are being used for AvailableValues, but for now this should fix the issue.

A unit test has been added to ensure the issue no appears in the future.