460 Dependency Resolver With GraphQL: Improvement
- Added performance improvement so that initially, only release packages are queried.

If RCs, betas or alphas are needed they will be separate queries for each package name. This has replaces one huge query with several small queries, that in total are much smaller.

Currently the repository does not support multiple packages in the graphql query, so the only solution is to split them into multiple queries. Consider doing multiple queries in parallel for improved percieved performance.