876 Uninstall Delete Retry: added delete retry to delete files while in used during install
In some unlikely cases, it seems like the file can be in use while uninstalling. The root cause for this was never found and it only happens intermittently on some PCs. Since it is probably an AV issue or such, let's retry and give the user a chance to close the document or application.