879 Parameters Performance Issue Fix
Added an optimization in the parameter member annotator that recursively collapses multiple layers of parameters and then only apply the logic on those root members.

This has a big effect for SweepRowMemberData as it iself is an IParameterMemberData, so whenever that gets annotated it always has at least two levels of indirection, which causes a lot of additional overhead.