Fix Resolver Issues: Fix
Under some (reproducible, but not understood) circumstances the resolver could find a result which did not contain the dependencies of a package.

While I was unable to establish exactly why this happens, I was able to add a check for it and rerun parts of the resolver and then adding those dependencies when needed.

This seems to fix the issue on those systems which showed the issue.