915 Unable To Delete Bug: Fix
For unknown reasons, a package can contain the same file multiple times. This is an issue in its own right. When this happens the file will get deleted twice, which causes and exception to occur on the second delete.

Apparantly a File.Delete can throw a DirectoryNotFoundException if the directory of the file being deleted does not exist.