961 Race-Condition in result table optimizer: Fix
Fixed the issue by adding an additional check in CreateOptimizedTable: if the peeked object from the work queue does not have the same proxy / result source, then it comes from another step and in this case we break the table in smaller chunks.

In general this should only rarely happen since it only occurs when steps publish compatible result tables in parallel.