922 improve usability of package list when no releases are available
We currently give a somewhat cryptic hint that there are no 'released' versions of a package,
but it's not really obivous that this means that there are **pre**-released versions.

From the logic, there was supposed to be a message hinting that **all** versions could
be shown with the  option, or by providing a prerelease specifier like

Instead of this teasing, I think we should just show the list of prereleases in the
case where a user tried to list the versions of a package with no releases.

Closes #922