980 add Required option to CommandLineArgumentAttribute
This adds the Required option to CommandLineArgumentAttribute

This allows authors of CLI actions to indicate that a named argument
is required. If the required argument is not provided, the CLI action
executor will give an appropriate error message, and fail with the
help text for the CLI action.

This should save CLI action authors to avoid a lot of boiler plate
that we ourselves duplicate in all of our plugins:

Code like this should no longer be necessary:
```csharp
if (string.IsNullOrWhiteSpace(SomeProp))
  throw new Exception($"{nameof(SomeProp)} is required");

```

Closes #980