991 add suggested corrections to package actions (#994)
* 991 add suggested corrections to package actions

This adds a UserInput with valid corrections if a user specifies
a package name that we don't know about. E.g:

```bash
$ tap package show opentap`
Package 'opentap' not found. Did you mean 'OpenTAP'?
0: 'Yes' (default)
1: 'No'
Please enter a number or name (Yes):
```

If there are multiple likely options:
```bash
$ tap package list 'editors'
Package 'editors' not found. Did you mean:
0: 'Editor' (default)
1: 'Editor X'
2: 'Editor CE'
3: 'Cancel'
Please enter a number or name (Editor):
```

The fuzzy matcher is not *that* smart, but can be improved in
a future release if there are glaring problems.

Unfortunately, this feature adds a noticeable delay, especially
if a slow repository is configured. This is because the feature adds a
round-trip to all repositories in all cases, except the case where the
user correctly types the name of a package which is available locally.

Closes #991

---------

Co-authored-by: StefanHolst <stefan.holst@keysight.com>
Co-authored-by: Asger <asger_iversen@keysight.com>