1051 bug resolving beta dependencies (#1077)
* 1075 Pacakge Create Circular Dependencies: Fix

Fixed the issue by checking if the added package dependency depends on the package itself.

Also added check that throws an exception if a circular dependency is added. This should only occur if some error happened.

* 1051 Bug Resolving Beta Dependencies: Fix

The issue was caused by a bug occuring if an rc was resolved before a beta. The logic for comparing beta and rc was reversed.

beta.CompareTo(rc) is < 0, not opposite.

Added additional fix, the problem was: A:beta -> B:beta -> C:beta -> D:beta

In this case we need to ensure that all the packages are cached in the graph of available packages.