1291 Parameterized mixin bug: Fix
Added IDynamicMember to MixinMemberData and a way for it to communicate that the mixin member has been disposed. This means that now the parameterization of the mixin gets removed if the mixin is removed.

Also added a fix in case the parameterized member is an embedded member data.

Added unit test to verify the fix.