1293 TestPlanReference With Mixins (#1295)
* 1293 TestPlanReference With Mixins: Fix

Fixed an issue with loading test plan references with mixins.

also added unit tests to verify the issue.

* 1292 Error Loading test plan with mixin.

Fixed the issue that test plans could not load with mixins.

Added unit test to verify the issue.

* 1291 Parameterized mixin bug: Fix

Added IDynamicMember to MixinMemberData and a way for it to communicate that the mixin member has been disposed. This means that now the parameterization of the mixin gets removed if the mixin is removed.

Also added a fix in case the parameterized member is an embedded member data.

Added unit test to verify the fix.