1329 Modify Mixin Not Working Fix (#1332)
This is hard to reproduce without installing the Expressions package.

The expressions wants to hide modify mixin if an expression is already assigned to that member. To do this, it has to get the member of the mixin member in the menu model. In this case the member was null, which meant that the expressions plugin just checked if any property of the step had an expression rather than the specific property.