Count actual number of commits since version change
This solves an issue where the version number of a branch can be reset in some circumstances after a merge.
This can in the following scenario:

a feature branch is created branching off from the current main branch
Some some commits are added to main
commits are added to feature branch
main is merged into the feature branch with a merge commit
the feature branch is fast-forwarded into main

Because the merge commit from merging main into the feature branch was fast forwarded back onto main without creating a new merge commit,
the gitversion calculation would now skip all commits from between that merge commit, and the first parent of that merge commit.
This would cause the prerelease counter to potentially rewind.

Instead, with this commit, we count the actual number of commits. This has a couple of side effects. For example, when a feature branch
is merged into main with a merge commit, the prerelease counter of the main branch will increment once for each commit that was merged, and once
for the merge commit. Previously, the prerelease counter would have only incremented once.

This is a bit unfortunate, but unavoidable. But at least the prerelease counter is now guaranteed to be monotonically increasing, except when force pushing to a detached head.