1092 accept piped input in ConsoleUserInput
This adds a new reading mode when Console.Stdin is redirected.
The existing method using Console.ReadKey fails when input is redirected.
The new method reads bytes directly from the stdin stream, and adds the
lines to a line buffer which is used when UserInput is requested.
This also allows other processes to proxy UserInput from an OpenTAP
subprocess, which previously would not have been possible without
emulating a terminal host.

Closes #1092