1368 Output Type Compatibility: Fix
Allow outputs to be assigned to more types of inputs. Some conversion is done between them. Sometimes this conversion might fail. This should result in a runtime error.

Sub classes are allowed to be assigned to parent classes, but not the other way around.

Unit tests were added to verify the behavior.

Added documentation about inputs and outputs in general.