1368 Output Type Compatibility: Fix (#1484)
* 1368 Output Type Compatibility: Fix

Allow outputs to be assigned to more types of inputs. Some conversion is done between them. Sometimes this conversion might fail. This should result in a runtime error.

Sub classes are allowed to be assigned to parent classes, but not the other way around.

Unit tests were added to verify the behavior.

Added documentation about inputs and outputs in general.

* 1486 Chained Outputs Fix:

- Fixed race condition occuring when the step is completed but the OnTestStepRunCompleted has not been pushed to the work queue yet. This happens about 1/1000.

- Fixed issue where chained outputs throws an timeout exception even though the outputs will become available eventually.