1545 mixins added twice bug (#1549)
* 1545 Mixins Added Twice Bug
Added unit test to verify the issue.

Fixed the issue by checking if an element has already been added with that name.

Also modified the mixins so that the element name is the same as the member name. This should be safe since we dont support two mixins with the same name anyway.