Set assembly version to gitversion during build
This adds a new target to the nuget package to set gitversion during
build instead of during package creation.

This is nice for a couple of reasons:

* we can remove the dependency on mono.cecil
* shorter package.xml files since version injection no longer needs to
be specified for all dlls
* The hash of a packaged DLL will match the hash of the input dll,
unless signed / obfuscated of course
* dlls will be versioned correctly in debug builds
* we no longer need to update pdb files when we update the version info,
since this is now done correctly by the compiler

I had to make GitVersionCalculator public. Let's consider if we want to
actually do this. It would not be a big deal if it was in the 'SDK'
package, but unfortunately it is in OpenTap.Package.