1656 No Output for isolated process: Fix (#1664)
After some investigation, it turned out that the error can occur with processes started with the -h argument. This is probably because they exist very quickly after printing the output, but in general probably any error can get lost.

To fix it I added waiting for the output writing tasks to finish before ending the process.