1824 Multi-Select Convert to Sequence: Fix
Added support for multi-selecting test plan reference and converting many to sequences with just a couple of clicks.

Added IMergedValueAnnotation to be able to show annotators when we are dealing with merged data.

Added IMergedMethodAnnotation for exactly the case where we can manually call the method in a combined way.