1846 Multi Level scope bug: Fix
Fixed the issue by setting the Parent to null when a test step is removed from a TestStepList.

This fixes the problem because it is now possible to correctly find the child step from a disconnected parent, if both the parent and child has been disconnected from the test plan.