Restrict collection serializer factory to only apply to explicit dotnet types
This fixes an issue causing the factory logic to apply to all kinds of
typedata implementations: ParameterMemberData, ExpandedMemberData, and
any plugin-defined member data. MemberData from such typedata should
have already been deserialized in the object that actually owns the
member, and thus no special handling is required.

This also allows removing the special-case for test plan parameters that
was added in #1853