Update version sort logic in resolver
This change ensures that when no prerelease is specified, more stable
versions are preferred over less stable versions. This is an extension
to the existing logic that any compatible release version is preferred
over a prerelease, unless the prerelease was explicitly requested. With
the new logic, we prefer release > rc > beta > alpha, regardless of
version number, if no version number is specified.