fix MemberValueAnnotation cache inconsistency
This fixes an issue when writing overlapping merged parameters. If a
member was marked as invalidated by e.g. calling 'Read', it would not
properly read back the underlying value on a subsequent Write() call.
This caused it to write back an outdated value to its source objects in
some scenarios.