serializer: defer property setter (#2093)
* serializer: defer property setter

This fixes an issue where parameter members are cloned before they are
fully deserialized if the parameterized member uses deferred setters.

This also introduces the concept of an ordering of deferred actions.
For example the TestPlanReference deserializer relied on deferring
actions within a defer statement to ensure it runs late in the defer
process, which was not compatible with the strategy of deferring setting
member values.

Also adds a unittest verifying the behavior

Closes #2084

* Verify behavior of new serializer defer

* serializer: Update outdated comment

* Update defer priority enum with named defer phases

This centralizes the order of deferred statements in the enum. This
should be clearer than some vague notion of ordering like 'early' and
'late'