uninstall: file deletion ensure same volume
Previous iterations attempted to uninstall by moving files to the temp
directory, but
File.Move has different semantics when moving files between storage
volumes (e.g. C:\ to D:\ drive)
In such cases, it will copy the file and then attempt to delete it
normally. This doesn't work
on Windows if the file is in use. To work around this limitation, we
uninstall files to a subdirectory of
the OpenTAP installation; this should ensure the file can always be
moved.