Package install rework 2 (#2119)
* upgrade: ensure tap.dll is updated

This adds logic to replace tap.dll in addition to tap.exe when upgrading
on Windows.

This also greatly simplifies Tap.Upgrader.exe by moving files instead of
deleting them if they are in use.

Closes #2120

* Add retry logic

* Simplify startup to stop using isolation

* Delete RestartManager

This is no longer needed with the new package install strategy

* Installer: Move files instead of deleting

This allows in-place updates on Windows

* main: Set isolation environment variable

This is needed for compatibility with old OpenTAP plugins which expect
to run isolated.

* Disable logic starting isolated processes

* Explicitly build tap

This is needed because OpenTap.Package cannot depend on tap (framework).
Previously Tap.Upgrader had this dependency

* Move package creation to tap.csproj

* Include hash for tap.exe

* Factor file renaming into class

* Re-add tap.upgrader noop binary

* move startup logic back to TapEntry

This ensures backwards compatibility with versions of OpenTAP which
cannot upgrade tap.exe in-place

* Undo removal of tap.upgrader

* uninstall: file deletion ensure same volume

Previous iterations attempted to uninstall by moving files to the temp
directory, but
File.Move has different semantics when moving files between storage
volumes (e.g. C:\ to D:\ drive)
In such cases, it will copy the file and then attempt to delete it
normally. This doesn't work
on Windows if the file is in use. To work around this limitation, we
uninstall files to a subdirectory of
the OpenTAP installation; this should ensure the file can always be
moved.

* install: reintroduce WaitForPackageFilesFree

* tap.csproj: stop linking shared project

* delete old uninstalls on UninstallContext creation

* set default build configuration to debug

* remove additional build invocations

* Delete RunIsolated implementation

* Added uninstall context logging and cleanup

Also added unit tests

various other cleanups.

* Minor cleanups

Fixed an issue that the Editor package cannot be upgraded after it is installed on Linux.

* Fix minor typo

---------

Co-authored-by: Rolf Madsen <rolf_madsen@keysight.com>