UserInput: respect LayoutMode for readonly items
This updates the UserInput implementation to respect the LayoutMode
property.

Members with LayoutMode set to Normal will now be prepended with
"<Name>: "

The text content for such items are also indented to the same level.
I opted to not implement this for writable properties because it looks
kind of silly with the "Please Enter <name> (default value):"
decorations.

Closes #2147