pluginsearcher: fix race conditions
* add locks where non-threadsafe internal data structures are modified
This fixes a race when TypeData.FromType() is being called from
different threads.
* clone PluginSearcher instead of mutating and invalidating
This resolves race conditions when TypeData.FromType() is being called
while searching. Now, such calls will either operate on the old
PluginSearcher instance, or it will block on the call to GetSearcher()
which ensures the current search is finished.
* clone old TypeData instead of mutating TypeData datastructures
This resolves strange behavior in calls to e.g.
TypeData.GetDerivedTypes() while TypeData.DerivedTypes is being
invalidated.